Add more checks for uneven width and height
Certain parts of the script do not play well with uneven dimensions. There was a check for an uneven height, but not an uneven width. This fixes it.

Both checks are also moved to after the resizing functions as SD and HD resizing usually result in even dimensions. If neither happens or if uneven dimensions still remain, the checks will account for that.